Logically separate plugins, assistants, and managers
